compute shader writing to texture, that is then read by the pixel shader

+++ rendering/dxUpdateTextureCL.h
@@ -38,7 +38,7 @@ namespace rendering
 			int signal,
 			std::string& errorMessage);
 
-		void GenerateTexData(std::vector<UINT8>& data, DXTexture* tex);
+		void GenerateTexData(std::vector<float>& data, DXTexture* tex);
 	public:
 	};
 }
